requirements.txt
itk-elastix
matplotlib
numpy
pandas
scikit-learn